FAST shooting!!! candle blinking, new amp setpoints
Fast shooting powered by fast alignment!!!! Turns out the problem all along with our alignment programs was that we were instantiating the SwerveRequest in the class and not in any method. We have to instantiate it in the initialize method of the command if it is actually going to do anything. Sucks that it *only* took most of the build season to figure that one out.